% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastrender}
\alias{fastrender}
\title{Fast but sloppy HTML rendering}
\usage{
fastrender(x, indent = 0L)
}
\arguments{
\item{x}{An htmltools tag object.}

\item{indent}{Initial indent level. Use a negative value for no indentation.}
}
\value{
An HTML string (marked as if returned from \link[htmltools]{HTML}).
}
\description{
Renders \pkg{htmltools} tag graphs into HTML. This is missing several
features of \link[htmltools]{renderTags}, including but not necessarily
limited to: singleton, head, and dependency handling; converting objects
to tags using \link[htmltools]{as.tags}; combining attributes with the
same name by space-delimiting them; and possibly others.
}

